requirements.txt
adjustText
gradio
matplotlib
numpy
Pillow
